refactor: simplify hallucination fallback logic by failing fast
